Update programme 6.py

+++ user-input/programme 6.py
@@ -2,10 +2,11 @@
 #Date = 1st of Semptember 2025
 #description = Program to request user name as input and return the same with a greeting 
 total_bill = input('Enter total Bill : ')
-total_bill = int(total_bill)
+total_bill = float(total_bill)
 
 no_diners = input('Number of diners : ')
-no_diners = int(no_diners)
+no_diners = float(no_diners)
 each_p = total_bill/no_diners
 print('Each person should pay',each_p)
+
